% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edges.R
\name{add_edges}
\alias{add_edges}
\title{Add edges to graph}
\usage{
add_edges(graph_id, edges, vertices, edge_src_col = 1,
  edge_dst_col = 2, edge_weight_col, edge_type_col, vertex_id_col = 1,
  auto_add_vertices = TRUE)
}
\arguments{
\item{graph_id}{the graph to which to add the edges and vertices. Should
be of type \code{chickenwire}.}

\item{edges}{a data.frame withe edges. Should contain at least two columns:
source vertex id's of the edge and destination id's. By default it is assumed 
that these are in the first two columns of the data.frame (see \code{edge_src_col}
and \code{edge_dst_col}).}

\item{vertices}{a data.frame with vertices (optional). By default it is assumed that
the first column contains the vertex id's (see \code{vertex_id_col}).}

\item{edge_src_col}{name or number of the column from \code{edges} containing the source 
vertex id's of the edges. Column should contain integer values.}

\item{edge_dst_col}{name or number of the column from \code{edges} containing the destination 
vertex id's of the edges. Column should contain integer values.}

\item{edge_weight_col}{name or number of the column from \code{edges} containing the weights 
of the edges. When omitted weights of 1 are assumed. Column should contain 
numeric values.}

\item{edge_type_col}{name or number of the column from \code{edges} containing the types 
of the edges. When omitted all edges have type 1. Should be of type integer.}

\item{vertex_id_col}{name or number of the column from \code{vertices} containing
the id's of the vertices. Should contain integer values.}

\item{auto_add_vertices}{automatically create vertices when an edge contains id's that
do not exist yet. When \code{vertices} is given this parameter is set to \code{FALSE}.}
}
\value{
Modifies \code{graph_id} and returns \code{graph_id}.
}
\description{
Add edges to graph
}
\details{
Note that the original \code{graph_id} is modified.
}
\examples{
# Create graph with three vertices and 5 edges between them.
g <- create_graph()
g <- add_edges(g, data.frame(src = c(1,1,2,2,3), dst = c(2,3,1,3,1)))
print(g)
delete_graph(g)

# Create graph with four vertices and 5 edges between them; one of the
# vertices does not have edges.
g <- create_graph()
g <- add_edges(g, data.frame(src = c(1,1,2,2,3), dst = c(2,3,1,3,1)),
  vertices = data.frame(id = c(1,2,3,4)))
print(g)
delete_graph(g)

}
