% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edges_raw.R
\name{add_edges_raw}
\alias{add_edges_raw}
\title{Raw/direct interface for adding edges to graph}
\usage{
add_edges_raw(graph_id, edges, edge_src_col = 1, edge_dst_col = 2,
  edge_weight_col, edge_type_col)
}
\arguments{
\item{graph_id}{the graph to which to add the edges and vertices. Should
be of type \code{chickenwire}.}

\item{edges}{a data.frame withe edges. Should contain at least two columns:
source vertex id's of the edge and destination id's. By default it is assumed 
that these are in the first two columns of the data.frame (see \code{edge_src_col}
and \code{edge_dst_col}).}

\item{edge_src_col}{name or number of the column from \code{edges} containing the source 
vertex id's of the edges. Column should contain integer values.}

\item{edge_dst_col}{name or number of the column from \code{edges} containing the destination 
vertex id's of the edges. Column should contain integer values.}

\item{edge_weight_col}{name or number of the column from \code{edges} containing the weights 
of the edges. When omitted weights of 1 are assumed. Column should contain 
numeric values.}

\item{edge_type_col}{name or number of the column from \code{edges} containing the types 
of the edges. When omitted all edges have type 1. Should be of type integer.}
}
\value{
Modifies \code{graph_id} and returns \code{graph_id}.
}
\description{
Adds edges to a graph. Assumes that vertices are numbered 0 and up. Does not
recode the source and destination id's of the edges. Does not add vertices
when these do not yet exists (this will generate an error; edges are added up 
until the error).
}
\details{
Note that the original \code{graph_id} is modified.
}
\examples{
# Create graph with three vertices and 5 edges between them.
g <- create_graph()
g <- add_vertices_raw(g, data.frame(id = c(0,1,2,3,4)))
g <- add_edges_raw(g, data.frame(src = c(0,0,1,1,2), dst = c(1,2,0,2,0)))
print(g)
delete_graph(g)

}
\seealso{
Is intended to be used together with \code{\link{add_vertices_raw}} to create a 
graph where the user takes care of correctly labelling edges and vertices.
}
