% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vertices.R
\name{add_vertices}
\alias{add_vertices}
\title{Add vertices to graph}
\usage{
add_vertices(graph_id, vertices, vertex_id_col = 1L)
}
\arguments{
\item{graph_id}{the graph to which to add the edges and vertices. Should
be of type \code{chickenwire}.}

\item{vertices}{a data.frame or vector with vertices. By default it is assumed that
the first column of the data.frame contains the vertex id's (see \code{vertex_id_col}). This
column, or the vector supplied should of type \code{numeric}, \code{character} or 
\code{factor}.}

\item{vertex_id_col}{name or number of the column from \code{vertices} containing
the id's of the vertices. Used when a data.frame is supplied for \code{vertices}.}
}
\value{
Modifies \code{graph_id} and returns \code{graph_id}.
}
\description{
Add vertices to graph
}
\details{
Note that the original \code{graph_id} is modified.
}
\examples{
# Create a graph with 4 vertices without edges
g <- create_graph()
g <- add_vertices(g, data.frame(id = c(1,2,3,4)))
print(g)
delete_graph(g)


}
