% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graph.R
\name{create_graph}
\alias{create_graph}
\title{Create empty graph}
\usage{
create_graph(directed = TRUE)
}
\arguments{
\item{directed}{create directed graph}
}
\value{
An object of type \code{chickenwire} which is a graph. In this case an empty
graph to which edges and vertices can be added using \code{\link{add_edges}} 
and \code{\link{add_vertices}}. See those functions for examples.
}
\description{
Create empty graph
}
\details{
Internally the graph is always stored as a directed graph. The option to 
specify that the graph is undirected has mainly effect when adding edges: 
when the edge A->B is added, the edge B->A is also added automatically.
}
