% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_all_graphs.R
\name{delete_all_graphs}
\alias{delete_all_graphs}
\title{Delete all graph}
\usage{
delete_all_graphs()
}
\value{
The memory belonging to all graphs is cleared and all \code{graph_id} are no longer
valid. Subsequent use of a \code{graph_id} will give an error.
}
\description{
Delete all graph
}
