% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_graph.R
\name{delete_graph}
\alias{delete_graph}
\title{Delete graph}
\usage{
delete_graph(graph_id)
}
\arguments{
\item{graph_id}{the graph to which to add the edges and vertices. Should
be of type \code{chickenwire}.}
}
\value{
The memory belonging to the graph is cleared and \code{graph_id} is no longer
valid. Subsequent use of \code{graph_id} will give an error.
}
\description{
Delete graph
}
