% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_directed.R
\name{is_directed}
\alias{is_directed}
\title{Get if graph is directed or not}
\usage{
is_directed(graph_id)
}
\arguments{
\item{graph_id}{the graph to which to add the edges and vertices. Should
be of type \code{chickenwire}.}
}
\value{
Boolean indicating whether the graph is directed or not.
}
\description{
Get if graph is directed or not
}
