% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_average.R
\name{local_average}
\alias{local_average}
\title{Local network averages}
\usage{
local_average(graph_id, vertex_values, vertex_weights = 1,
  alpha = 0.85, nstep_max = 200, nworkers = available_cores(),
  precision = 1e-05)
}
\arguments{
\item{graph_id}{the graph to which to add the edges and vertices. Should
be of type \code{chickenwire}.}

\item{vertex_values}{a vector with vertex values. Can be numeric, factor or
character values.}

\item{vertex_weights}{a vector with vertex weights. Should be numeric and 
have the same length as \code{vertex_values}. Optional.}

\item{alpha}{probability of continuing the random walk (1-restart 
probability).}

\item{nstep_max}{maximum number of iterations.}

\item{nworkers}{number of threads to use during the calculation.}

\item{precision}{when updates in the local averages are smaller than 
\code{precision} the algorithm is stopped.}
}
\value{
A data.frame. The rows correspond to the vertices.
}
\description{
Local network averages
}
