% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{nedges}
\alias{nedges}
\title{Get number of edges in graph}
\usage{
nedges(graph_id)
}
\arguments{
\item{graph_id}{the graph to which to add the edges and vertices. Should
be of type \code{chickenwire}.}
}
\value{
An integer with the number of edges.
}
\description{
Get number of edges in graph
}
