% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{nvertices}
\alias{nvertices}
\title{Get number of vertices in graph}
\usage{
nvertices(graph_id)
}
\arguments{
\item{graph_id}{the graph to which to add the edges and vertices. Should
be of type \code{chickenwire}.}
}
\value{
An integer with the number of vertices.
}
\description{
Get number of vertices in graph
}
