% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ssi}
\alias{ssi}
\title{Example network simulating pink and purple neighbourhoods}
\format{An object of class \code{list} of length 2.}
\usage{
ssi
}
\description{
Example network simulating pink and purple neighbourhoods
}
\references{
\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{datasets}
