% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices.R
\name{vertices}
\alias{vertices}
\title{Get the vertices of a graph}
\usage{
vertices(graph_id)
}
\arguments{
\item{graph_id}{the graph to which to add the edges and vertices. Should
be of type \code{chickenwire}.}
}
\value{
Data.frame with vertices.
}
\description{
Get the vertices of a graph
}
