% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_average.R
\name{local_average}
\alias{local_average}
\title{Local network averages}
\usage{
local_average(graph_id, vertex_values, vertex_weights = 1)
}
\arguments{
\item{graph_id}{the graph to which to add the edges and vertices. Should
be of type \code{chickenwire}.}

\item{vertex_values}{a vector with vertex values. Can be numeric, factor or
character values.}

\item{vertex_weights}{a vector with vertex weights. Should be numeric and 
have the same length as \code{vertex_values}. Optional.}
}
\value{
A data.frame. This first column contains the vertex id's.
}
\description{
Local network averages
}
